move lint_errors_parser
